cvd_controller/requirements.txt
# CVD Controller – Python dependencies
# Install with: pip install -r requirements.txt

PyQt6>=6.6.0
pyserial>=3.5
pyqtgraph>=0.13.0
